Generate all xtz-shots website from one metadata file
To use:

`python snapshot_jekyll.py`

This generates a file in _data/snapshot_jekyll_data.json based on the
global metadata files `snapshots.json` containing all artifacts for all
networks.

This file is used by jekyll to generate the entire xtz-shots website.

I am using a new plugin "jekyll-datapage-gen" since jekyll has no
default capability to generate pages.

snapshot_jekyll.py should not be merged. Instead, the contents should
move to the python file in tezos-k8s zip-and-upload.

When zip-and-upload is done generating the global metadata file, it
should then generate this jekyll page and put it in a special bucket
(not where the artifacts are). This bucket will be exposed as top-level
page (xtz-shots.io)

The new page for latest artifacts is now `xtz-shots.io/<network_name>`
instead of `<network_name>.xtz-shots.io`. A redirect should be
configured in the existing buckets so users don't complain.

This POC is not using remote theme, but I assume remote-theme should
work out-of-the-box (just add the plugin to the config)

Also in snapshotEngine:

* there is no longer any need to download `snapshot.md`, it should come
  as part of the remote theme,
* all the code generating json files with jq can be removed, everything
  comes from the unique metadata file now.