metadata guide + branding
Add a metadata section in the doc with a jq example.

Remove snapshot duration from doc since it's no longer current.

Add a mention of metadata in the "all snapshots" page.

Change the title of the regular snapshot page to "Most recent snapshots
for..."

Provide a link to older snapshots in the most recent snapshots page.

Change the layout of the most recent snapshot page to utilize all page
width (no more oxhead banner on the top right).

add oxheadalpha icon, change links to oxhead website

change favicon to oxheadalpha